+++ CHANGELOG.md
@@ -2,6 +2,11 @@ The upgrade instructions are available at [Oro documentation website](https://do
 
 The current file describes significant changes in the code that may affect the upgrade of your customizations.
 
+### Removed
+#### FrontendBundle
+* `orofrontend/default/js/app/views/input-widget/checkbox` was removed; use pure CSS checkbox customization instead.
+* `orofrontend/default/js/app/views/input-widget/checkbox-radio` was removed; use pure CSS radio button customization instead.
+
 ## 5.1.0-alpha.1 (2022-05-31)
 [Show detailed list of changes](incompatibilities-5-1-alpha.md)